backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3 
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
whitenoise==6.7.0       
gunicorn==21.2.0       